+++ fissure/utils/alert_sender.py
@@ -31,31 +31,29 @@ def _alert_sender(cmd: str, c2: Connection, identifier: str, sensor_node_id: any
     logger : logging.Logger
         Sensor node logger
     """
-    # run command
+    # Run command
     proc = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, stdin=subprocess.PIPE, bufsize=1, universal_newlines=True, shell=True)
 
-    # monitor for alerts
-    print("ALERT SENDER!!!!!!!!!111111111111")
+    # Monitor for alerts
     stop = False
     while not stop and proc.poll() is None:
         for proc_text in proc.stdout:
 
-            # replace any gps fields in text
+            # Replace any gps fields in text
             proc_text = proc_text % gps_position
 
             try:
                 data = json.loads(proc_text)
 
                 if not data.__class__ is dict:
-                    # not within the scope of alert messaging; log as info
+                    # Not within the scope of alert messaging; log as info
                     logger.info(str(data))
 
                 elif not 'msg' in list(data.keys()):
-                    # not within the scope of alert messaging; log as info
+                    # Not within the scope of alert messaging; log as info
                     logger.info(str(data))
 
                 elif data.get('msg') == 'alert':
-                    print("AAAAAAAAAAAAAAALLLLLLLLLLLLLEEEEEEEEEERT")
                     if network_type == "IP":
                         PARAMETERS = {
                             "sensor_node_id": sensor_node_id,
@@ -91,23 +89,24 @@ def _alert_sender(cmd: str, c2: Connection, identifier: str, sensor_node_id: any
                             "time": data.get('time'),
                             "remarks": data.get('remarks'),
                         }
-                        msg = {
-                            fissure.comms.MessageFields.IDENTIFIER: identifier,
-                            fissure.comms.MessageFields.MESSAGE_NAME: "takPlot",
-                            fissure.comms.MessageFields.PARAMETERS: PARAMETERS,
-                        }
-                        asyncio.run(hiprfisr_socket.send_msg(fissure.comms.MessageTypes.COMMANDS, msg))
+
+                        if data.get('remarks') == "GPS UPDATE":
+                            msg = {
+                                fissure.comms.MessageFields.IDENTIFIER: identifier,
+                                fissure.comms.MessageFields.MESSAGE_NAME: "takPlotGpsUpdate",
+                                fissure.comms.MessageFields.PARAMETERS: PARAMETERS,
+                            }
+                            asyncio.run(hiprfisr_socket.send_msg(fissure.comms.MessageTypes.COMMANDS, msg))                        
+
+                        else:
+                            msg = {
+                                fissure.comms.MessageFields.IDENTIFIER: identifier,
+                                fissure.comms.MessageFields.MESSAGE_NAME: "takPlot",
+                                fissure.comms.MessageFields.PARAMETERS: PARAMETERS,
+                            }
+                            asyncio.run(hiprfisr_socket.send_msg(fissure.comms.MessageTypes.COMMANDS, msg))
                         
                     elif network_type == "Meshtastic":
-                        # PARAMETERS = {
-                            # "uid": data.get('uid'),
-                            # "lat": data.get('lat'),
-                            # "lon": data.get('lon'),
-                            # "alt": data.get('alt'),
-                            # "time": data.get('time'),
-                            # "remarks": data.get('remarks'),
-                        # }
-                        #PARAMETERS = {"msg": [data.get('uid'), data.get('lat'), data.get('lon'), data.get('alt'), data.get('time'), data.get('remarks')]}
                         PARAMETERS = {
                             "msg": [
                                 data.get('uid'),
@@ -118,13 +117,23 @@ def _alert_sender(cmd: str, c2: Connection, identifier: str, sensor_node_id: any
                                 data.get('remarks')[:20] if data.get('remarks') else None
                             ]
                         }
-                        msg = {
-                            fissure.comms.MessageFields.SOURCE: identifier,
-                            fissure.comms.MessageFields.DESTINATION: fissure.comms.Identifiers.HIPRFISR_LT,
-                            fissure.comms.MessageFields.MESSAGE_NAME: "takPlotLT",
-                            fissure.comms.MessageFields.PARAMETERS: PARAMETERS,
-                        }
+
+                        if data.get('remarks') == "GPS UPDATE":
+                            msg = {
+                                fissure.comms.MessageFields.SOURCE: identifier,
+                                fissure.comms.MessageFields.DESTINATION: fissure.comms.Identifiers.HIPRFISR_LT,
+                                fissure.comms.MessageFields.MESSAGE_NAME: "takPlotGpsUpdateLT",
+                                fissure.comms.MessageFields.PARAMETERS: PARAMETERS,
+                            }
+                        else:
+                            msg = {
+                                fissure.comms.MessageFields.SOURCE: identifier,
+                                fissure.comms.MessageFields.DESTINATION: fissure.comms.Identifiers.HIPRFISR_LT,
+                                fissure.comms.MessageFields.MESSAGE_NAME: "takPlotLT",
+                                fissure.comms.MessageFields.PARAMETERS: PARAMETERS,
+                            }
                         asyncio.run(hiprfisr_socket.send_msg(fissure.comms.MessageTypes.COMMANDS, msg))
+
                 elif data.get('msg') == 'exploit':
                     if network_type == "IP":
                         PARAMETERS = {
@@ -144,15 +153,6 @@ def _alert_sender(cmd: str, c2: Connection, identifier: str, sensor_node_id: any
                         asyncio.run(hiprfisr_socket.send_msg(fissure.comms.MessageTypes.COMMANDS, msg))
 
                     elif network_type == "Meshtastic":
-                        # PARAMETERS = {
-                            # "sensor_node_id": sensor_node_id,
-                            # "protocol": data.get('protocol'),
-                            # "modulation": data.get('modulation'),
-                            # "hardware": data.get('hardware'),
-                            # "type": data.get('type'),
-                            # "attack": data.get('attack'),
-                            # "variables": data.get('variables'),
-                        # }
                         PARAMETERS = {
                             "msg": [
                                 sensor_node_id,
@@ -174,15 +174,15 @@ def _alert_sender(cmd: str, c2: Connection, identifier: str, sensor_node_id: any
                         asyncio.run(hiprfisr_socket.send_msg(fissure.comms.MessageTypes.COMMANDS, msg))
 
             except json.decoder.JSONDecodeError:
-                # not a valid json string
+                # Not a valid json string
                 logger.info(str(proc_text))
 
-            if c2.poll(): # check for messages on c2 comms
-                msg = c2.recv() # get message
+            if c2.poll(): # Check for messages on c2 comms
+                msg = c2.recv()  # Get message
                 if msg == 'QUIT':
-                    stop = True # stop alert monitoring
-                    break # break out of for loop
-    proc.wait() # wait for process to end
+                    stop = True  # Stop alert monitoring
+                    break
+    proc.wait()  # Wait for process to end
 
 class alertSender(object):
     def __init__(self, cmd: str, identifier: str, sensor_node_id: any, hiprfisr_socket: fissure.comms.Server, gps_position: dict, logger: logging.Logger, network_type: str):