+++ fissure/utils/codes.py
@@ -19,6 +19,9 @@ MESSAGE_CODE_MAP = {
     "I1": "alertReturnLT",
     "J1": "takPlotLT",
     "K1": "exploitLT",
+    "L1": "takPlotGpsUpdateLT",
+    "M1": "gpsBeaconEnableMeshtasticLT",
+    "M2": "gpsBeaconDisableMeshtasticLT",
 }
 
 # Reverse mapping for sending messages